Create mf2fs.py
mf2fs.py: my files to folder structure. 
I needed a script to move a huge amount of picture and video files into an existing folderstructure based which was based on Y\M\D paths. This was the result.
It did the job and is now extended to use it as a management script to check existing cloud drives for file duplicates.